+++ raiden-ts/src/services/actions.ts
@@ -2,7 +2,7 @@
 import * as t from 'io-ts';
 
 import { createAction, ActionType, createAsyncAction } from '../utils/actions';
-import { Address, UInt, Signed } from '../utils/types';
+import { Address, UInt, Signed, Hash } from '../utils/types';
 import { Paths, PFS, IOU } from './types';
 
 const PathId = t.type({
@@ -45,3 +45,38 @@ export interface iouClear extends ActionType<typeof iouClear> {}
 
 export const udcDeposited = createAction('udc/deposited', UInt(32));
 export interface udcDeposited extends ActionType<typeof udcDeposited> {}
+
+const UdcWithdrawId = t.type({
+  amount: UInt(32),
+});
+
+export const udcWithdraw = createAsyncAction(
+  UdcWithdrawId,
+  'udc/withdraw/request',
+  'udc/withdraw/success',
+  'udc/withdraw/failure',
+  t.undefined,
+  t.intersection([
+    t.type({ block: t.number }),
+    t.partial({ txHash: Hash, txBlock: t.number, confirmed: t.union([t.undefined, t.boolean]) }),
+  ]),
+);
+
+export namespace udcWithdraw {
+  export interface request extends ActionType<typeof udcWithdraw.request> {}
+  export interface success extends ActionType<typeof udcWithdraw.success> {}
+  export interface failure extends ActionType<typeof udcWithdraw.failure> {}
+}
+
+export const udcWithdrawn = createAction(
+  'udc/withdrawn',
+  t.type({
+    withdrawal: UInt(32),
+    txHash: Hash,
+    txBlock: t.number,
+    confirmed: t.union([t.undefined, t.boolean]),
+  }),
+  UdcWithdrawId,
+);
+
+export interface udcWithdrawn extends ActionType<typeof udcWithdrawn> {}