+++ raiden-ts/tests/e2e/raiden.spec.ts
@@ -1,5 +1,5 @@
 import { timer } from 'rxjs';
-import { first, filter, takeUntil } from 'rxjs/operators';
+import { first, filter, takeUntil, take, toArray } from 'rxjs/operators';
 import { Zero, MaxUint256 } from 'ethers/constants';
 import { parseEther, parseUnits, bigNumberify, BigNumber, keccak256, Network } from 'ethers/utils';
 
@@ -28,7 +28,7 @@ import { makeInitialState, RaidenState } from 'raiden-ts/state';
 import { raidenShutdown, ConfirmableAction } from 'raiden-ts/actions';
 import { newBlock, tokenMonitored } from 'raiden-ts/channels/actions';
 import { ChannelState } from 'raiden-ts/channels/state';
-import { Storage, Secret, Address } from 'raiden-ts/utils/types';
+import { Storage, Secret, Address, UInt } from 'raiden-ts/utils/types';
 import { isActionOf } from 'raiden-ts/utils/actions';
 import { ContractsInfo } from 'raiden-ts/types';
 import { PartialRaidenConfig } from 'raiden-ts/config';
@@ -39,6 +39,8 @@ import { losslessStringify } from 'raiden-ts/utils/data';
 import { ServiceRegistryFactory } from 'raiden-ts/contracts/ServiceRegistryFactory';
 import { ErrorCodes } from 'raiden-ts/utils/error';
 import { channelKey } from 'raiden-ts/channels/utils';
+import { confirmationBlocks } from '../unit/fixtures';
+import { udcWithdrawn } from 'raiden-ts/services/actions';
 
 describe('Raiden', () => {
   const provider = new TestProvider();
@@ -1190,7 +1192,7 @@ describe('Raiden', () => {
     });
   });
 
-  describe('depositToUDC', () => {
+  describe('UDC', () => {
     test('deposit 0 tokens', async () => {
       expect.assertions(1);
       await expect(raiden.depositToUDC(0)).rejects.toThrow('positive');
@@ -1205,6 +1207,72 @@ describe('Raiden', () => {
       expect.assertions(1);
       await raiden.mint(await raiden.userDepositTokenAddress(), 10);
       await expect(raiden.depositToUDC(10)).resolves.toMatch(/^0x[0-9a-fA-F]{64}$/);
+    });
+
+    test('withdraw success', async () => {
+      expect.assertions(5);
+      const deposit = bigNumberify(100) as UInt<32>;
+      const withdraw = bigNumberify(80) as UInt<32>;
+      await raiden.mint(await raiden.userDepositTokenAddress(), deposit);
+      await expect(raiden.depositToUDC(deposit)).resolves.toMatch(/^0x[0-9a-fA-F]{64}$/);
+      await expect(raiden.getUDCCapacity()).resolves.toEqual(deposit);
+      await expect(raiden.planUdcWithdraw(withdraw)).resolves.toMatch(/^0x[0-9a-fA-F]{64}$/);
+
+      const result = raiden.action$.pipe(filter(udcWithdrawn.is), take(2), toArray()).toPromise();
+      await provider.mine(100 + confirmationBlocks * 2);
+
+      await expect(raiden.getUDCCapacity()).resolves.toEqual(deposit.sub(withdraw));
+      await expect(result).resolves.toEqual([
+        udcWithdrawn(expect.objectContaining({ withdrawal: withdraw, confirmed: undefined }), {
+          amount: withdraw,
+        }),
+        udcWithdrawn(expect.objectContaining({ withdrawal: withdraw, confirmed: true }), {
+          amount: withdraw,
+        }),
+      ]);
+    });
+
+    test('withdraw success with restart', async () => {
+      expect.assertions(5);
+      const deposit = bigNumberify(100) as UInt<32>;
+      const withdraw = bigNumberify(80) as UInt<32>;
+      await raiden.mint(await raiden.userDepositTokenAddress(), deposit);
+      await expect(raiden.depositToUDC(deposit)).resolves.toMatch(/^0x[0-9a-fA-F]{64}$/);
+      await expect(raiden.getUDCCapacity()).resolves.toEqual(deposit);
+      await expect(raiden.planUdcWithdraw(withdraw)).resolves.toMatch(/^0x[0-9a-fA-F]{64}$/);
+
+      // stop and restart node
+      raiden.stop();
+      await provider.mine(confirmationBlocks);
+      raiden = await createRaiden(0, storage);
+      raiden.start();
+
+      const result = raiden.action$.pipe(filter(udcWithdrawn.is), take(2), toArray()).toPromise();
+      await provider.mine(100 + confirmationBlocks);
+
+      await expect(raiden.getUDCCapacity()).resolves.toEqual(deposit.sub(withdraw));
+      await expect(result).resolves.toEqual([
+        udcWithdrawn(expect.objectContaining({ withdrawal: withdraw, confirmed: undefined }), {
+          amount: withdraw,
+        }),
+        udcWithdrawn(expect.objectContaining({ withdrawal: withdraw, confirmed: true }), {
+          amount: withdraw,
+        }),
+      ]);
+    });
+
+    test('withdraw failure zero amount', async () => {
+      expect.assertions(1);
+      await expect(raiden.planUdcWithdraw(0)).rejects.toThrow(
+        'The planned withdraw amount has to be greater than zero.',
+      );
+    });
+
+    test('withdraw failure zero balance', async () => {
+      expect.assertions(1);
+      await expect(raiden.planUdcWithdraw(10)).rejects.toThrow(
+        'The planned withdraw amount exceeds the total amount available for withdrawing.',
+      );
     });
   });